Use an extensible how-to and examples section
Per review comments, tidy up the README references to the how-tos and
configuration examples, and use a title / format to which more docs can
be easily added.

In the process, add a preliminary version of an example setup for French.
Currently a copy of the extra-dics example/how-to but will become a
unique document.

Signed-off-by: Daniel F. Dickinson <dfdpublic@wildtechgarden.ca>